topdown model, moderately works

+++ det_rnn/base/_stimulus.py
@@ -23,7 +23,7 @@ class Stimulus(object):
         mask            = self._gen_mask()
         return {'neural_input'  : neural_input.astype(np.float32),
                 'stimulus_ori'  : stimulus_ori,
-                'desired_output': desired_output,
+                'desired_output': desired_output.astype(np.float32),
                 'mask'          : mask}
 
     def _gen_stimseq(self):
